% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_anchor_positions.R
\name{cluster_anchor_positions}
\alias{cluster_anchor_positions}
\title{Summarise integration anchors}
\usage{
cluster_anchor_positions(
  anchor_positions,
  min_score = 0,
  max_score = 1,
  cluster_threshold = 5
)
}
\arguments{
\item{anchor_positions}{A raw set of anchor positions from get_positions_for_anchors}

\item{min_score}{The minimum score for the anchors to use}

\item{cluster_threshold}{The euclidean clustering threshold to use - higher gives fewer clusters}
}
\value{
A tibble with grouped and summarised anchor positions
}
\description{
Summarise integration anchors
}
