% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimplot_cluster_interactions.R
\name{dimplot_cluster_interactions}
\alias{dimplot_cluster_interactions}
\title{Dimensionality reduction plot with anchored clusters}
\usage{
dimplot_cluster_interactions(
  cluster_anchors,
  min_score = 0,
  clusters1 = "seurat_clusters",
  clusters2 = "seurat_clusters",
  reduction = "umap",
  min_percent = 10
)
}
\arguments{
\item{cluster_anchors}{Output of FindIntegrationAnchors on clustered data}

\item{min_score}{Minimum anchor score to use}

\item{clusters1}{Cluster metadata name in dataset 1}

\item{clusters2}{Cluster metadata name in dataset 2}

\item{reduction}{Name of reduction to use}

\item{min_percent}{Minimum percent anchor linkage to show in plot}
}
\value{
ggplot of dimension plot coloured by cluster showing linkages
}
\description{
Dimensionality reduction plot with anchored clusters
}
