% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_postions_for_anchors.R
\name{get_positions_for_anchors}
\alias{get_positions_for_anchors}
\title{Convert a list of integration anchors into pairwise positions in a dimensional reduction}
\usage{
get_positions_for_anchors(anchors, min_score = 0, reduction = "umap")
}
\arguments{
\item{anchors}{An output from FindIntegrationAnchors}

\item{min_score}{The minimum score for anchor strength to report}

\item{reduction}{The reduction to use for the position extraction}
}
\value{
A tibble containing pairs of cell IDs along with their xy coordinates
}
\description{
Convert a list of integration anchors into pairwise positions in a dimensional reduction
}
