% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_interaction_matrix.R
\name{plot_cluster_interaction_matrix}
\alias{plot_cluster_interaction_matrix}
\title{Plot a cluster level interaction matrix for anchors}
\usage{
plot_cluster_interaction_matrix(
  cluster_anchors,
  min_score = 0,
  clusters1 = "seurat_clusters",
  clusters2 = "seurat_clusters"
)
}
\arguments{
\item{cluster_anchors}{Output of FindIntegrationAnchors}
}
\value{
heatmap of cluster percentage interaction
}
\description{
Plot a cluster level interaction matrix for anchors
}
