% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_integration_anchors.R
\name{plot_integration_anchors}
\alias{plot_integration_anchors}
\title{Plot integration anchors onto a dimensionality reduction plot}
\usage{
plot_integration_anchors(
  split_objects,
  anchor_positions,
  reduction = "umap",
  min_score = 0,
  max_score = 1
)
}
\arguments{
\item{split_objects}{The split Seurat object list}

\item{anchor_positions}{The output of get_positions_for_anchors}

\item{reduction}{The name of the dimension reduction to use}

\item{min_score}{The minimum anchor score to use}
}
\value{
A ggplot showing the anchors overlaid on the dimension plot
}
\description{
Plot integration anchors onto a dimensionality reduction plot
}
